back-end/requirements.txt
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
bcrypt==4.1.2
blinker==1.7.0
cachelib==0.10.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
Flask-Session==0.5.0
google-api-core==2.15.0
google-auth==2.26.2
google-cloud-aiplatform==1.39.0
google-cloud-bigquery==3.16.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.11.0
google-cloud-storage==2.14.0
google-cloud-vision==3.5.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
numpy==1.26.3
openai==1.7.2
packaging==23.2
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
rsa==4.9
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
vertexai==0.0.1
Werkzeug==3.0.1
